Create IUnitOfWork interface and UnitOfWork class

+++ frontend/src/components/Admin/Booking/index.jsx
@@ -39,6 +39,8 @@ const Booking = () => {
         deleteBooking: hasPermissionDelete,
     } = useCheckPermission();
 
+    console.log('Selected Booking', selectedBookings);
+
     // For deleting selected bookings
     useEffect(() => {
         const deleteAllBookings = async () => {
@@ -46,7 +48,6 @@ const Booking = () => {
                 // Create payload for deletion (id list)
                 const payload = deleteAll.payload.map((bookingDelete) => ({ id: bookingDelete.id }));
                 const response = await axios.delete('http://localhost:5058/booking', { data: payload });
-                console.log(response);
                 if (response?.status === 200) {
                     const data = response?.data?.updatedBookings?.$values.map((booking) => {
                         booking.statusName =
@@ -113,7 +114,6 @@ const Booking = () => {
             const url = 'http://localhost:5058/booking';
             const headers = { headers: { 'Content-Type': 'application/json' } };
             const response = await axios.get(url, headers);
-            console.log('Bookings', response);
             if (response?.status === 200) {
                 const data = response?.data?.$values.map((booking) => {
                     booking.statusName =
@@ -152,7 +152,6 @@ const Booking = () => {
         try {
             const url = 'http://localhost:5058/booking';
             const response = await axios.delete(`${url}/${payload}`);
-            console.log(response);
             if (response?.status === 200) {
                 showToast(response?.data?.obj?.message || response?.data?.message, 'success');
                 setBookings((prev) => prev.filter((booking) => booking.id !== payload));
@@ -173,21 +172,15 @@ const Booking = () => {
         }
     };
 
-    const handleAddClicked = () => {
-        setShowPanel('add');
-        setSelectedBooking(null);
-    };
-
     const handleRowClicked = useCallback(async (event) => {
         const { id } = event;
         try {
             const url = 'http://localhost:5058/booking';
             const headers = { headers: { 'Content-Type': 'application/json' } };
             const response = await axios.get(`${url}/${id}`, headers);
-            console.log(response);
             if (response?.status === 200) {
                 setShowPanel('see');
-                setSelectedBooking(response?.data?.obj);
+                setSelectedBooking(response?.data);
             }
         } catch (error) {
             console.error('Error fetching user details:', error);
@@ -220,12 +213,10 @@ const Booking = () => {
     };
 
     const handleStatusChange = async (bookingId, statusCode) => {
-        console.log('Payload: ', bookingId, statusCode);
         try {
             const url = 'http://localhost:5058/booking/status';
             const headers = { headers: { 'Content-Type': 'application/json' } };
             const response = await axios.put(`${url}/${bookingId}`, statusCode, headers);
-            console.log(response);
             if (response?.status === 200) {
                 fetchBookings();
                 hideContext();
@@ -283,8 +274,6 @@ const Booking = () => {
             selector: (row) => row.actions,
         });
     }
-
-    console.log(menuVisible);
 
     const data = searchedBookings?.map((booking, index) => ({
         id: booking?.id,
@@ -381,17 +370,7 @@ const Booking = () => {
             {ToastContainer}
 
             <div className="d-flex align-items-center justify-content-between w-full py-4">
-                {deleteAll.count === 0 ? (
-                    hasPermissionCreate ? (
-                        <FiPlus
-                            size={30}
-                            className="p-1 rounded-2 text-white secondary-bg-color cursor-pointer"
-                            onClick={handleAddClicked}
-                        />
-                    ) : (
-                        <></>
-                    )
-                ) : hasPermissionDelete ? (
+                {deleteAll.count !== 0 && hasPermissionDelete ? (
                     <BsTrash
                         size={30}
                         className="p-1 rounded-2 text-white secondary-bg-color cursor-pointer"
